Fix easy pong wrapper
Previously, the wrapper always resets as soon as someone scores. This is problematic because agent always opens an episode, leaving it not being able to handle scenarios where the environment player kicks off a game.

This update prevents environment from resetting an episode when someone scores. In other words, now the agent plays the full episode as a series of one-score games, as opposed to only play the first score of each episode.